reorg read index
